requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pybullet>=3.2.5
pyyaml>=6.0
attrs>=21.0.0
pytest>=6.0.0
websockets>=10.0
asyncio-mqtt>=0.11.0
